+++ opennmt/utils/beam_search.py
@@ -388,7 +388,7 @@ def beam_search(symbols_to_logits_fn,
     logits = tf.reshape(flat_logits, [batch_size, beam_size, -1])
 
     # Convert logits to normalized log probs
-    candidate_log_probs = _log_prob_from_logits(logits)
+    candidate_log_probs = _log_prob_from_logits(tf.to_float(logits))
 
     # Multiply the probabilities by the current probabilities of the beam.
     # (batch_size, beam_size, vocab_size) + (batch_size, beam_size, 1)
